+++ src/Contract/ClientLoaderInterface.php
@@ -2,8 +2,8 @@
 
 namespace Heptacom\AdminOpenAuth\Contract;
 
-use Heptacom\AdminOpenAuth\Component\Contract\ClientInterface;
 use Heptacom\AdminOpenAuth\Exception\LoadClientException;
+use Heptacom\OpenAuth\Client\Contract\ClientContract;
 use Shopware\Core\Framework\Context;
 
 interface ClientLoaderInterface
@@ -11,7 +11,7 @@ interface ClientLoaderInterface
     /**
      * @throws LoadClientException
      */
-    public function load(string $clientId, Context $context): ClientInterface;
+    public function load(string $clientId, Context $context): ClientContract;
 
     public function create(string $providerKey, Context $context): string;
 }